Remove/replace bloq seeds to improve fungibility
Fungibility is an essential attribute of good money.  
Companies conducting blockchain analysis should not play any trusted role within any money protocol, since their business is breaching the privacy of the users of that money, or in other words: their very business is built upon the notion that they are not to be trusted by the users of the money.  
  
I propose the removal and or replacement of bloq.